resolving issues
New folder structure exists and is clean.
Imports resolve to the new strategy location.
No diagnostics reported in the moved engine and strategy files.
