+++ strategies/__init__.py
@@ -1,0 +1,3 @@
+from .trading_strategy import get_trading_strategy
+
+__all__ = ["get_trading_strategy"]
